+++ Assets/Scripts/EnemyHealth.cs
@@ -2,11 +2,11 @@ using UnityEngine;
 
 public class EnemyHealth : MonoBehaviour
 {
-    [SerializeField] private int _health = 5;
+    [SerializeField] private float _health = 5;
 
     //[SerializeField] private int _healthMax;
 
-    public void TakeDamage(int damage) {
+    public void TakeDamage(float damage) {
         if (damage > 0) {
             _health -= damage;
             if (_health <= 0) {
